Adiciona validações de exclusão em entidades relacionadas

+++ src/main/java/projeto/biblioteca/backend/service/AutorService.java
@@ -11,12 +11,14 @@ import projeto.biblioteca.backend.dto.AutorRequestDto;
 import projeto.biblioteca.backend.dto.AutorResponseDto;
 import projeto.biblioteca.backend.models.Autor;
 import projeto.biblioteca.backend.repository.AutorRepository;
+import projeto.biblioteca.backend.repository.LivroRepository;
 
 @Service
 @RequiredArgsConstructor
 public class AutorService {
   
   private final AutorRepository autorRepository;
+  private final LivroRepository livroRepository;
 
   public List<AutorResponseDto> listarAutores() {
     return autorRepository.findAll()
@@ -38,6 +40,10 @@ public class AutorService {
 
     if(!autorRepository.existsById(id)) {
       throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Autor não encontrado com ID: " + id);
+    }
+
+    if(livroRepository.existsByAutorId(id)) {
+      throw new ResponseStatusException(HttpStatus.CONFLICT, "O autor possui livros cadastrados e não pode ser excluído");
     }
 
     autorRepository.deleteById(id);
